# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aura — Social Energy Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InteractionInteractionType = typeof InteractionInteractionType[keyof typeof InteractionInteractionType];


export const InteractionInteractionType = {
  work_meeting: 'work_meeting',
  friend_hangout: 'friend_hangout',
  family_call: 'family_call',
  party: 'party',
  texting: 'texting',
  date: 'date',
  group_event: 'group_event',
  solo_activity: 'solo_activity',
  other: 'other',
} as const;

export interface Interaction {
  id: number;
  personName: string;
  interactionType: InteractionInteractionType;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyBefore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyAfter: number;
  energyDelta?: number;
  durationMinutes: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type InteractionInputInteractionType = typeof InteractionInputInteractionType[keyof typeof InteractionInputInteractionType];


export const InteractionInputInteractionType = {
  work_meeting: 'work_meeting',
  friend_hangout: 'friend_hangout',
  family_call: 'family_call',
  party: 'party',
  texting: 'texting',
  date: 'date',
  group_event: 'group_event',
  solo_activity: 'solo_activity',
  other: 'other',
} as const;

export interface InteractionInput {
  /** @minLength 1 */
  personName: string;
  interactionType: InteractionInputInteractionType;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyBefore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyAfter: number;
  /** @minimum 1 */
  durationMinutes: number;
  notes?: string;
}

export type InteractionUpdateInteractionType = typeof InteractionUpdateInteractionType[keyof typeof InteractionUpdateInteractionType];


export const InteractionUpdateInteractionType = {
  work_meeting: 'work_meeting',
  friend_hangout: 'friend_hangout',
  family_call: 'family_call',
  party: 'party',
  texting: 'texting',
  date: 'date',
  group_event: 'group_event',
  solo_activity: 'solo_activity',
  other: 'other',
} as const;

export interface InteractionUpdate {
  /** @minLength 1 */
  personName?: string;
  interactionType?: InteractionUpdateInteractionType;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyBefore?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyAfter?: number;
  /** @minimum 1 */
  durationMinutes?: number;
  /** @nullable */
  notes?: string | null;
}

export type InteractionSummaryBurnoutLevel = typeof InteractionSummaryBurnoutLevel[keyof typeof InteractionSummaryBurnoutLevel];


export const InteractionSummaryBurnoutLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

export interface InteractionSummary {
  /** @nullable */
  todayEnergyScore: number | null;
  /** @nullable */
  weeklyAvgDelta: number | null;
  burnoutLevel: InteractionSummaryBurnoutLevel;
  totalInteractionsThisWeek: number;
  /** @nullable */
  mostDrainingType: string | null;
  /** @nullable */
  mostEnergizingType: string | null;
}

export interface PersonEnergy {
  personName: string;
  avgEnergyDelta: number;
  totalInteractions: number;
  /** @nullable */
  lastInteractionAt: string | null;
}

export interface EnergyTrendDay {
  date: string;
  netEnergyDelta: number;
  interactionCount: number;
}

export type InsightCategory = typeof InsightCategory[keyof typeof InsightCategory];


export const InsightCategory = {
  pattern: 'pattern',
  burnout_warning: 'burnout_warning',
  recovery_suggestion: 'recovery_suggestion',
  social_tip: 'social_tip',
  safe_person: 'safe_person',
  circle_analysis: 'circle_analysis',
} as const;

export interface Insight {
  id: number;
  content: string;
  category: InsightCategory;
  createdAt: string;
}

export interface SafePerson {
  personName: string;
  avgEnergyDelta: number;
  totalInteractions: number;
  /** 0-100 score indicating how consistently this person energizes you */
  safetyScore: number;
  /** AI-generated one-line description of this relationship */
  summary: string;
}

export type CircleConnectionClassification = typeof CircleConnectionClassification[keyof typeof CircleConnectionClassification];


export const CircleConnectionClassification = {
  energizing: 'energizing',
  neutral: 'neutral',
  draining: 'draining',
} as const;

export interface CircleConnection {
  personName: string;
  avgEnergyDelta: number;
  totalInteractions: number;
  classification: CircleConnectionClassification;
  /** Short AI-generated insight about this person */
  insight: string;
}

export type SocialCircleAnalysisInteractionBalance = {
  energizingCount: number;
  neutralCount: number;
  drainingCount: number;
  energizingPercent: number;
  drainingPercent: number;
};

export interface SocialCircleAnalysis {
  safePeople: SafePerson[];
  connections: CircleConnection[];
  interactionBalance: SocialCircleAnalysisInteractionBalance;
  /** 0-100 score for overall social circle health */
  overallHealthScore: number;
  /** AI-generated one-paragraph summary of the user's social circle */
  summary: string;
}