# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aura — Social Energy Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /insights/generate:
    post:
      operationId: generateInsights
      tags: [insights]
      summary: Trigger AI analysis and generate new insights
      responses:
        "200":
          description: Newly generated insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
        "400":
          description: Not enough data
...
components:
  schemas:
...
    Insight:
      type: object
      required: [id, content, category, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        category:
          type: string
          enum: [pattern, burnout_warning, recovery_suggestion, social_tip, safe_person, circle_analysis]
        createdAt:
          type: string
          format: date-time
...